[Update] 🔥 회원 정보 가져오기

+++ GoGuMa/src/main/java/com/ggm/goguma/service/cart/CartService.java
@@ -6,7 +6,7 @@ import com.ggm.goguma.dto.cart.CartItemDTO;
 
 public interface CartService {
 
-	List<CartItemDTO> getCartList(int memberId) throws Exception;
+	List<CartItemDTO> getCartList(long memberId) throws Exception;
 
 	void addCartCount(long cartId) throws Exception;
