+++ GoGuMa/src/main/java/com/ggm/goguma/service/cart/CartServiceImpl.java
@@ -20,7 +20,7 @@ public class CartServiceImpl implements CartService{
 	private CartMapper cartMapper;
 	
 	@Override
-	public List<CartItemDTO> getCartList(int memberId) throws Exception {
+	public List<CartItemDTO> getCartList(long memberId) throws Exception {
 		return cartMapper.getCartList(memberId);
 	}
